(wip) .types into the iframe'ed Hyperswitch fields; clicks Add Card.
So this is where I was prior but since it's instrumented I can pass the output back to the LLM or an agent and hopefully get a smarter than me loop going. I should probably adopt a KISS approach and use bog standard input fields and see if that goes to confirmPaymnet a-ok first, and skip the branch of it working with UnifiedPayment (?); or I just dynamically add that element; that's easier than some of the iframe busting techinques and test automation I just snuck in here.